Reformat services and refactor ProductivityEmailService
- Move RangeDates to dtos package
- Add convertToRangeDates private method in ProductivityEmailService
- Rename sendProductivityEmail to sendingRunnable
- Refactor askForRange
- Rename askForRange to sendProductivityEmail
